+++ client/src/container/container.ts
@@ -1,22 +1,24 @@
 import { Container } from "inversify";
 
 import { App } from "../app/App";
-import { CameraService } from "../services/CameraService";
+import { RotatingCubeScene } from "../scenes/rotating-cube/RotatingCubeScene";
 import { LayoutService } from "../services/LayoutService";
-import { SceneService } from "../services/SceneService";
-import { Service3D } from "../services/Service3D";
+import { WebGLRenderer } from "../services/WebGLRenderer";
+import type { Scene } from "../types/Scene";
 import { SERVICE_TYPES } from "./serviceTypes";
 
 export function createContainer(): Container {
   const container = new Container();
 
   container.bind<App>(SERVICE_TYPES.App).to(App).inSingletonScope();
-  container.bind<CameraService>(SERVICE_TYPES.CameraService).to(CameraService).inSingletonScope();
   container.bind<LayoutService>(SERVICE_TYPES.LayoutService).to(LayoutService).inSingletonScope();
-  container.bind<SceneService>(SERVICE_TYPES.SceneService).to(SceneService).inSingletonScope();
   container
-    .bind<Service3D>(SERVICE_TYPES.Service3D)
-    .to(Service3D)
+    .bind<Scene>(SERVICE_TYPES.Scene)
+    .to(RotatingCubeScene)
+    .inSingletonScope();
+  container
+    .bind<WebGLRenderer>(SERVICE_TYPES.WebGLRenderer)
+    .to(WebGLRenderer)
     .inSingletonScope();
 
   return container;