+++ docs/changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v1.0.1
+
+- express -> fastify
+- CameraService removed
+- SceneService -> RotatingCubeScene
+- Service3D -> WebGLRenderer
+- Scene and SceneComponent moved to `src/types`
+- scene components moved to `src/scene-components`
+
 ## v1.0.0
 
 - static server